Move heroName to dedicated file
